Spread out duration.rs into a module
Update copyright
