+++ pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["maturin>=0.13,<0.14"]
+requires = ["maturin>=1.1,<1.2"]
 build-backend = "maturin"
 
 [project]
